Commit #3
- Added Water
- Added Swimming
- Added Lava
- Added Drowning
- Added Crouching
- Added Checkpoints
- Added CoinTracker
- Added Brittle Blocks
- Added Ultimate Skill Choice in UI
- Changed Dojo Wall Sprite
- Changed Dojo Ground Sprite
- Rescaled main UI

+++ Assets/Scripts/Level/DamageableScript.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class DamageableScript : MonoBehaviour
 {
-    public int hitPoints;
+    public float hitPoints;
     public bool isBurning;
 
-    virtual public void GetDamaged(int damage, Collider2D instigator)
+    virtual public void GetDamaged(float damage, Collider2D instigator)
     {
 
     }
